perf: optimize affix
