+++ packages/zarm/src/affix/index.tsx
@@ -1,4 +1,12 @@
-import React, { useState, useEffect, forwardRef, useRef, useCallback, CSSProperties } from 'react';
+import React, {
+  useState,
+  useEffect,
+  forwardRef,
+  useRef,
+  useCallback,
+  CSSProperties,
+  HTMLAttributes,
+} from 'react';
 import classnames from 'classnames';
 import { canUseDOM } from '../utils/dom';
 import { AffixProps as BaseAffixProps } from './interface';
@@ -12,10 +20,9 @@ export interface AffixStates {
   height: number;
 }
 
-export interface AffixProps extends BaseAffixProps {
+export interface AffixProps extends BaseAffixProps, Omit<HTMLAttributes<HTMLElement>, 'onChange'> {
   prefixCls?: string;
   className?: string;
-  children?: React.ReactNode;
 }
 
 const DEFAULT_SCROLL_CONTAINER = canUseDOM ? window : undefined;
