+++ packages/zarm/src/affix/interface.ts
@@ -1,0 +1,8 @@
+import { ContainerType } from '../utils/dom';
+
+export interface AffixProps {
+  offsetTop?: number;
+  offsetBottom?: number;
+  scrollContainer?: ContainerType;
+  onChange?(affixed: boolean): void;
+}
